+++ handler/exchange.go
@@ -18,7 +18,7 @@ func (exchange) Types(_ string, ctx rpc.Context) (resp response) {
 }
 
 // List ...
-func (exchange) List(size, page int64, order string, ctx rpc.Context) (resp response) {
+func (exchange) List(size, page int, order string, ctx rpc.Context) (resp response) {
 	username := ctx.GetString("username")
 	if username == "" {
 		resp.Message = constant.ErrAuthorizationError
