+++ model/algorithm.go
@@ -20,7 +20,7 @@ type Algorithm struct {
 	Traders []Trader `gorm:"-" json:"traders"`
 }
 
-func (user User) ListAlgorithm(size, page int64, order string) (total int64, algorithms []Algorithm, err error) {
+func (user User) ListAlgorithm(size, page int, order string) (total int64, algorithms []Algorithm, err error) {
 	_, users, err := user.ListUser(-1, 1, "id")
 	if err != nil {
 		return
