+++ model/exchange.go
@@ -21,7 +21,7 @@ type Exchange struct {
 }
 
 // ListExchange ...
-func (user User) ListExchange(size, page int64, order string) (total int64, exchanges []Exchange, err error) {
+func (user User) ListExchange(size, page int, order string) (total int64, exchanges []Exchange, err error) {
 	_, users, err := user.ListUser(-1, 1, "id")
 	if err != nil {
 		return
